kubeflow_women_clothing_ecom_reviews/pipeline/components/model_training/requirements.txt
pyyaml==5.4.1
pandas==1.1.5
scikit-learn==0.24.1
xgboost==1.3.3
lightgbm==3.2.0
# for local testing
# kfp==1.4.0